// backend/src/auth/enums/error.messages.enum.ts
export enum ErrorMessagesEnum {
    USER_ALREADY_EXISTS = 'User already exists',
    USER_NOT_FOUND = 'User does not exist',
    PASSWORD_INCORRECT = 'Password is incorrect',
    EMAIL_NOT_VERIFIED = 'Your email is not verified',
    USER_INACTIVE = 'Your account is inactive',
    ACCESS_DENIED = 'Access Denied',
}